Uc12 throw user friendly error if employee id not found

+++ src/main/java/com/bridgelabz/employeepayrollservice/service/EmployeePayrollService.java
@@ -1,9 +1,7 @@
 package com.bridgelabz.employeepayrollservice.service;
-/**
- * import classes
- */
 
 import com.bridgelabz.employeepayrollservice.dto.EmployeeDTO;
+import com.bridgelabz.employeepayrollservice.exception.EmployeePayrollException;
 import com.bridgelabz.employeepayrollservice.model.Employee;
 import com.bridgelabz.employeepayrollservice.repository.EmployeePayrollRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,85 +10,61 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 
-/**
- * @Service
- *    @Service annotation is used in your service layer and annotates classes that perform service tasks
- */
+
 @Service
+
+//Created EmployeePayrollService class to serve api calls done by controller layer
 public class EmployeePayrollService implements IEmployeePayrollService {
+
+    //Autowired EmployeePayrollRepository interface to inject its dependency here
     @Autowired
     EmployeePayrollRepository repository;
 
-    public String getMessage(String name) {
-        return "Welcome To Employee Program " + name;
-    }
-
-    public String postMessage(EmployeeDTO employee) {
-        return "Hello Employee " + employee.getFirstName() + "" + employee.getLastName() + "!";
-    }
-
-    public String putMessage(String name) {
-        return "Hey Dude , " + name;
-    }
-
     public String getWelcome() {
-        return "Welcome to Employee Payroll App.....!";
+        return "Welcome to Employee Payroll !!!";
     }
 
-    /**
-     * accepts the employee data in the form of EmployeeDTO and stores it in DB
-     * @param employee - employee data
-     * @return - accepted employee information in JSON format
-     */
-    @Override
-    public Employee postDataToRepo(EmployeeDTO employee) {
-        Employee newEmployee = new Employee(employee);
+    public Employee postDataToRepo(EmployeeDTO employeeDTO) {
+        Employee newEmployee = new Employee(employeeDTO);
         repository.save(newEmployee);
         return newEmployee;
     }
 
-    /**
-     *
-     * @return - return list of employee information from DB
-     */
-    @Override
     public List<Employee> getAllData() {
         List<Employee> list = repository.findAll();
         return list;
     }
 
-    /**
-     * @param id - represents employee id
-     * @return - employee information with same empId
-     */
-    @Override
-    public Optional<Employee> getDataById(Integer id) {
+    public Employee getDataById(Integer id) {
         Optional<Employee> newEmployee = repository.findById(id);
-        return newEmployee;
+        if (newEmployee.isPresent()) {
+            return newEmployee.get();
+        } else throw new EmployeePayrollException("Employee id not found");
     }
 
-    /**
-     * accepts the employee data in the form of EmployeePayrollDTO and
-     * updates the employee having same empId from database
-     * @param id - employee id
-     * @param employeeDTO - represents object of EmployeePayrollDTO class
-     * @return - updated employee information in JSON format
-     */
     public Employee updateDataById(Integer id, EmployeeDTO employeeDTO) {
-        Employee newEmployee = new Employee(id, employeeDTO);
-        repository.save(newEmployee);
-        return newEmployee;
+        Optional<Employee> newEmployee = repository.findById(id);
+        if (newEmployee.isPresent()) {
+            Employee employee = new Employee(id, employeeDTO);
+            repository.save(employee);
+            return employee;
+        } else {
+            throw new EmployeePayrollException("Employee Not found");
+        }
     }
 
-    /**
-     * accepts the empId and deletes the data of that employee from DB
-     * @param id - represents employee id
-     * @return - empId and Acknowledgment message
-     */
     public String deleteDataById(Integer id) {
-        repository.deleteById(id);
-        return "Employee with unique ID:" + id + " got deleted";
+        Optional<Employee> newEmployee = repository.findById(id);
+        if (newEmployee.isPresent()) {
+            repository.deleteById(id);
+        } else {
+            throw new EmployeePayrollException("Employee Details not found");
+        }
+      return null;
     }
-
-
 }
+
+
+
+
+
